gateway-api: Move GAMMA watch handlers to watchhandlers module
This commit moves GAMMA watch handlers to the new watchhandlers
module.
